% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_small_area_core_outputs.R
\name{arrange_small_area_core_outputs}
\alias{arrange_small_area_core_outputs}
\title{Arrange outputs from the \code{small_area_led_core} function into inputs for the
\code{output_small_area_projection} function}
\usage{
arrange_small_area_core_outputs(
  projection,
  popn_backseries,
  dwelling_trajectory,
  first_proj_yr,
  final_proj_yr
)
}
\arguments{
\item{projection}{A list. The output list from the \code{small_area_led_core} function}

\item{popn_backseries}{A dataframe. The small area population backseries}

\item{dwelling_trajectory}{A dataframe. The small area dwelling trajectory by year
and geographic aggregation (ward or msoa)}

\item{first_proj_yr}{Numeric. First projection year}

\item{final_proj_yr}{Numeric. Last projection year}
}
\value{
A list where each element is a data frame containing data for each year
  of the projection and the backseries.
}
\description{
The \code{small_area_led_core} outputs a list for each year of the projection.
Each list is a list of components. This function rearrages those outputs into
a list of dataframes of components.
}
