% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ward_config_to_msoa_config.R
\name{convert_ward_config_to_msoa_config}
\alias{convert_ward_config_to_msoa_config}
\title{A simple function to convert a ward model config to an msoa model config}
\usage{
convert_ward_config_to_msoa_config(ward_config_list)
}
\arguments{
\item{ward_config_list}{A small area model ward config list}
}
\value{
An msoa config list for the small area model
}
\description{
Goes through the ward config list and replaces the string \code{ward} with
the string \code{msoa}. Also changes the path of the small area to district
lookup file.
}
