% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_small_area_projection.R
\name{output_small_area_projection}
\alias{output_small_area_projection}
\title{Output the results of the housing-led model}
\usage{
output_small_area_projection(
  projection,
  output_dir,
  projection_type,
  births,
  deaths,
  first_proj_yr,
  lookup
)
}
\arguments{
\item{projection}{A list. The output from the run_housing_led_model function}

\item{output_dir}{A string. The directory in which to save the model output files}

\item{projection_type}{A string. Is the projection at ward or MSOA level? Used in
naming the output files only}

\item{births}{A dataframe. The projeted births output from the small area model}

\item{deaths}{A dataframe. The projeted deaths output from the small area model}

\item{first_proj_yr}{Numeric. The first year of the projection}

\item{lookup}{A dataframe. A gss_code to name lookup for the modelled areas}
}
\description{
Save RDS and CSV files to a specified folder path. Outputs include population,
components of change (including backseries data) and the input dwelling and
household trajectories.
}
