% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_small_area_model.R
\name{run_small_area_model}
\alias{run_small_area_model}
\title{Run a small area model config file to produce a population projection}
\usage{
run_small_area_model(config_list)
}
\arguments{
\item{config_list}{A List. A housing-led model configuration list.}
}
\value{
A list of projected components
}
\description{
Read in, validate and manage model input data and then run the \code{small_area_core}
function to produce a population projection as specified in the config list.
}
