% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_area_core.R
\name{small_area_core}
\alias{small_area_core}
\title{Small area model core: calculate components of change and next year's
population}
\usage{
small_area_core(
  start_population,
  births,
  deaths,
  communal_est_popn,
  out_migration_rates,
  in_migration_characteristics,
  popn_constraint,
  birth_constraint,
  death_constraint,
  fertility_rates,
  mortality_rates,
  last_data_yr,
  dwellings,
  adults_per_dwelling,
  projection_year,
  small_area_to_district
)
}
\arguments{
\item{start_population}{Data frame. Initial population at small area
resolution.}

\item{births}{Data frame. Historical births backseries at small area
resolution. Used when \code{projection_year} is \code{last_data_yr} or
earlier.}

\item{deaths}{Data frame. Historical deaths backseries at small area
resolution. Used when \code{projection_year} is \code{last_data_yr} or
earlier.}

\item{communal_est_popn}{Data frame. Static communal establishment population
at small area resolution.}

\item{out_migration_rates}{Data frame. Static outmigration rates at small
area resolution.}

\item{in_migration_characteristics}{Data frame. Static inmigration rates at
small area resolution. Rates should give the number of people by age and
sex that migrate to an area for each adult (over 18) that migrates to an
area (that is, the rates will sum to a bit more than 1 for each geography).}

\item{popn_constraint}{Data frame. Population by borough to constrain the
output to.}

\item{birth_constraint}{Data frame. Births by borough to constrain the output
to.}

\item{death_constraint}{Data frame. Deaths by borough to constrain the output
to.}

\item{fertility_rates}{NULL or data frame. Fertility rates at small area
resolution. Used when \code{projection_year} is after
\code{last_data_yr}.}

\item{mortality_rates}{NULL or data frame. Mortality rates at small area
resolution. Used when \code{projection_year} is after
\code{last_data_yr}.}

\item{last_data_yr}{Integer. Year at which the model switches from using
historical births and deaths from \code{births} and \code{deaths} to
projected rates from \code{fertility_rates} and \code{mortality_rates}.}

\item{dwellings}{Data frame. Number of dwellings at small area resolution.}

\item{adults_per_dwelling}{Data frame. Ratio of adults to dwellings at small
area resolution.}

\item{projection_year}{Integer. Uh, the projection year.}

\item{small_area_to_district}{Data frame. A lookup from ward to borough codes.}
}
\description{
Small area model core: calculate components of change and next year's
population
}
